refactor(App): ♻️ move function GenerateLocaleFiles to writer
