+++ utils/file.go
@@ -1,11 +1,9 @@
 package utils
 
 import (
+	"fmt"
 	"os"
-
-	"github.com/appsynth-org/transit/reader"
-	android "github.com/appsynth-org/transit/writer/android"
-	ios "github.com/appsynth-org/transit/writer/ios"
+	"path/filepath"
 )
 
 func GetFileExtension(platform string) string {
@@ -38,39 +36,13 @@ func HasTranslation() bool {
 	return true
 }
 
-func GenerateLocaleFiles(groups []reader.LocalizeGroup) {
-	iosWriterEN := ios.NewDocument("en")
-	iosWriterTH := ios.NewDocument("th")
-
-	androidWriterEN := android.NewDocument("en")
-	androidWriterTH := android.NewDocument("th")
-
-	for _, group := range groups {
-		iosWriterEN.WriteComment(group.GroupName)
-		iosWriterTH.WriteComment(group.GroupName)
-
-		androidWriterEN.WriteComment(group.GroupName)
-		androidWriterTH.WriteComment(group.GroupName)
-
-		for _, translation := range group.Keys {
-			// Skip Android empty key
-			if len(translation.AndroidKey) > 0 {
-				androidWriterEN.WriteAttribute(translation.AndroidKey, translation.Translation.En)
-				androidWriterTH.WriteAttribute(translation.AndroidKey, translation.Translation.Th)
-			}
-
-			// Skip iOS empty key
-			if (len(translation.IosKey)) > 0 {
-				iosWriterEN.WriteAttribute(translation.IosKey, translation.Translation.En)
-				iosWriterTH.WriteAttribute(translation.IosKey, translation.Translation.Th)
-			}
-
+func CreateDirectoryIfNotExist(path string) error {
+	dir := filepath.Dir(path)
+	if _, err := os.Stat(dir); os.IsNotExist(err) {
+		err := os.MkdirAll(dir, 0755)
+		if err != nil {
+			return fmt.Errorf("error creating directory: %w", err)
 		}
 	}
-
-	iosWriterEN.Close()
-	iosWriterTH.Close()
-
-	androidWriterEN.Close()
-	androidWriterTH.Close()
+	return nil
 }
