+++ writer/ios/ios.go
@@ -4,7 +4,8 @@ import (
 	"fmt"
 	"log"
 	"os"
-	"path/filepath"
+
+	"github.com/appsynth-org/transit/utils"
 )
 
 type Document struct {
@@ -30,20 +31,9 @@ func (d *Document) Close() {
 	d.File.Close()
 }
 
-func createDirectoryIfNotExist(path string) error {
-	dir := filepath.Dir(path)
-	if _, err := os.Stat(dir); os.IsNotExist(err) {
-		err := os.MkdirAll(dir, 0755)
-		if err != nil {
-			return fmt.Errorf("error creating directory: %w", err)
-		}
-	}
-	return nil
-}
-
 func NewDocument(lang string) *Document {
 	filePath := fmt.Sprintf("../../output/iOS/%s.strings", lang)
-	createDirectoryIfNotExist(filePath)
+	utils.CreateDirectoryIfNotExist(filePath)
 
 	file, err := os.OpenFile(filePath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0644)
 	if err != nil {
